refactor: extract podman container management to separate module
Move all podman and container lifecycle functions (`run_podman`,
`launch_container`, `kill_container`, `configure_container`, etc.)
from `browsers_api_router.py` to a new `podman_browsers.py` module.

This separates infrastructure concerns from API endpoints and WebSocket
proxy logic, making both modules more focused.
